Implement proper negative constant support in IR and ARM64 code generation
- Change IR Const struct to use signed int64_t instead of uint64_t for proper negative value representation
- Update Lifter to generate signed constants from RISC-V immediate values, removing unnecessary casts
- Enhance InstructionSelector to detect negative constants (-1 to -65536) and generate MOVN instructions
- Simplify ARM64 encoder MOV handling by removing complex negative detection logic
- Add proper MOVN encoding support in ARM64 encoder for negative immediates
- Remove unused executeProgram and executeWithDBT methods from BinaryTranslator
- Move instruction analysis logging from executeProgram to executeFunction for better debugging
- Update unit tests to handle signed constant values
- Add clear error handling for constants outside supported range (throw for >16-bit values)

This fixes the "MOV immediate value too large" error for RISC-V negative immediates like -32, -20, -24
by properly representing them as signed values and using ARM64 MOVN instructions where ~imm16 produces
the desired negative result.

+++ lib/Lifter.cpp
@@ -142,15 +142,13 @@ void Lifter::liftSingleInstruction(const riscv::Instruction &inst) {
 
   // Upper immediate instructions
   case riscv::Instruction::Opcode::LUI: {
-    ir::ValueId imm = createConstant(
-        ir::Type::i64, static_cast<uint64_t>(inst.getImmediate(1)) << 12);
+    ir::ValueId imm = createConstant(ir::Type::i64, inst.getImmediate(1) << 12);
     setRegisterValue(inst.getRegister(0), imm);
     break;
   }
   case riscv::Instruction::Opcode::AUIPC: {
     ir::ValueId pc = createConstant(ir::Type::i64, inst.address);
-    ir::ValueId imm = createConstant(
-        ir::Type::i64, static_cast<uint64_t>(inst.getImmediate(1)) << 12);
+    ir::ValueId imm = createConstant(ir::Type::i64, inst.getImmediate(1) << 12);
     ir::ValueId result =
         createBinaryOp(ir::BinaryOpcode::Add, ir::Type::i64, pc, imm);
     setRegisterValue(inst.getRegister(0), result);
@@ -177,7 +175,7 @@ void Lifter::setRegisterValue(uint32_t regNum, ir::ValueId valueId) {
   }
 }
 
-ir::ValueId Lifter::createConstant(ir::Type type, uint64_t value) {
+ir::ValueId Lifter::createConstant(ir::Type type, int64_t value) {
   return addInstruction(ir::Const{type, value});
 }
